Squashed 'modules/core/dependency/python-ihm/' changes from 0a6e31129e..1d6f0599b1
1d6f0599b1 Fix typos
b190922c24 Fix typos
614bf5e84e Fix hyperlinks
1e2e35a308 Merge pull request ihmwg/python-ihm#122 from grandrea/cross_linkers_ag
1828c76e41 Match names in doc against Python code
1c53df70b9 Match crosslink names with dictionary
da722510f7 Drop testing with Python 2.7
779b38d71f formatting
d0a038af5d define dsbu, PhoX and tbu-PhoX. Missing smiles and inchi on PhoX and tbu-PhoX as they don't seem to exist yet.
1022dbf56e define photoL
363f48172d define photoL
1805462213 define photoL
a05ef17b53 Fill in entity for Residues in asyms
57c2d25f40 Remove db_name parameter from DatabaseLocation
660ccf8fb1 Make sure that input and output are not the same file

git-subtree-dir: modules/core/dependency/python-ihm
git-subtree-split: 1d6f0599b180267a13d64a0dd1ddcc2a1910d5d8

+++ modules/core/dependency/python-ihm/util/make-mmcif.py
@@ -22,6 +22,7 @@ import ihm.dumper
 import ihm.model
 import ihm.protocol
 import sys
+import os
 
 
 def add_ihm_info(s):
@@ -61,6 +62,10 @@ if len(sys.argv) > 2:
     out_fname = sys.argv[2]
 else:
     out_fname = 'output.cif'
+
+if (os.path.exists(fname) and os.path.exists(out_fname)
+        and os.path.samefile(fname, out_fname)):
+    raise ValueError("Input and output are the same file")
 
 with open(fname) as fh:
     with open(out_fname, 'w') as fhout:
